+++ project/src/arenashooter/engine/animation/tracks/AnimTrackVec2f.java
@@ -9,15 +9,15 @@ import arenashooter.engine.math.Vec2f;
  */
 public class AnimTrackVec2f extends AnimTrack<Vec2f> {
 
-	public AnimTrackVec2f(Map<Float, Vec2f> keyframes) {
+	public AnimTrackVec2f(Map<Double, Vec2f> keyframes) {
 		super(keyframes);
 	}
 
 	@Override
-	public Vec2f valueAt(float time) {
+	public Vec2f valueAt(double time) {
 		int prev = prevKeyframe(time);
 		int next = nextKeyframe(time);
-		float f = times[next]-times[prev];
+		double f = times[next]-times[prev];
 		if(f != 0)
 			f = (time-times[prev])/f;
 		return Vec2f.lerp((Vec2f)values[prev], (Vec2f)values[next], f);
