+++ project/src/arenashooter/engine/animation/tracks/AnimTrackVec3f.java
@@ -9,15 +9,15 @@ import arenashooter.engine.math.Vec3f;
  */
 public class AnimTrackVec3f extends AnimTrack<Vec3f> {
 
-	public AnimTrackVec3f(Map<Float, Vec3f> keyframes) {
+	public AnimTrackVec3f(Map<Double, Vec3f> keyframes) {
 		super(keyframes);
 	}
 
 	@Override
-	public Vec3f valueAt(float time) {
+	public Vec3f valueAt(double time) {
 		int prev = prevKeyframe(time);
 		int next = nextKeyframe(time);
-		float f = times[next]-times[prev];
+		double f = times[next]-times[prev];
 		if(f != 0)
 			f = (time-times[prev])/f;
 		return Vec3f.lerp((Vec3f)values[prev], (Vec3f)values[next], f);
